refactor(navigation): split tab and activity components
